Don't half-ass next_gen-ing
_funcs is terribly generic too
